+++ src/main/java/com/lucianaugusto/ebanxassignment/event/operation/BalanceInfo.java
@@ -2,9 +2,7 @@ package com.lucianaugusto.ebanxassignment.event.operation;
 
 import com.lucianaugusto.ebanxassignment.balance.model.Balance;
 
-import java.math.BigDecimal;
-
-public record BalanceInfo(String accountNumber, BigDecimal amount) {
+public record BalanceInfo(String accountNumber, Integer amount) {
 
     public BalanceInfo(Balance balance) {
         this(balance.getAccountNumber(), balance.getAmount());
